+++ tests/src/metadata_tests.cpp
@@ -5,11 +5,16 @@
 
 namespace databento::tests {
 TEST(MetadataTests, TestDatasetConditionDetailToString) {
-  const DatasetConditionDetail target{"2022-11-10", DatasetCondition::Available,
-                                      "2023-03-01"};
-  ASSERT_EQ(
-      ToString(target),
+  const DatasetConditionDetail available{
+      "2022-11-10", DatasetCondition::Available, "2023-03-01"};
+  EXPECT_EQ(
+      ToString(available),
       R"(DatasetConditionDetail { date = "2022-11-10", condition = available, last_modified_date = "2023-03-01" })");
+  const DatasetConditionDetail missing{
+      "2022-11-11", DatasetCondition::Missing, {}};
+  EXPECT_EQ(
+      ToString(missing),
+      R"(DatasetConditionDetail { date = "2022-11-11", condition = missing, last_modified_date = nullopt })");
 }
 
 TEST(MetadataTests, TestDatasetRangeToString) {